Updating README
I thought it would be fun to add a bit of a logo for Unite. I wasn't
sure what the best mascot for the plugin would be, but figured it would
be hard to beat a dragon. I went with a dragon, one who's body is in the
shape of a "U". I thought it came out pretty well, especially for a quick
effort.

Added an animated gif screen cast of some basic unite functionality.
This seemed like a better way to get the basic concepts of the plugin
across quickly.

Rearranged a bit of the content to group resources together in one spot.